mobile-app/backend/requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
Pillow>=10.0.0
supabase>=2.0.0
gunicorn>=21.0.0
